refactor: directories in sync with packages
