+++ build.sbt
@@ -138,11 +138,11 @@ lazy val `scafi-mp-api` = crossProject(JSPlatform, JVMPlatform, NativePlatform)
 
 lazy val `scafi-integration` = project
   .in(file("scafi-integration"))
-  .dependsOn(`scafi-distributed`.jvm % "test->test", `scafi-mp-api`.js % "compile->compile;test->test")
+  .dependsOn(`scafi-distributed`.jvm)
   .settings(commonDependencies)
   .settings(
-    publish / skip := true,
     fork := false,
+    publish / skip := true,
     Test / test := (Test / test).dependsOn(`scafi-mp-api`.js / Compile / fullLinkJS).tag(ExclusiveTestTag).value,
   )
